+++ resources/views/categories/categories.blade.php
@@ -3,7 +3,7 @@
 @section('content')
         <!--Middle Content-->
 
-         <a href="AddCategory"><button type="button" class="button2">Add Category</button> </a>
+         <a href="AddCategory"><button type="button" class="button9">Add Category</button> </a>
 
 
 
@@ -11,22 +11,8 @@
               <div class="col-md-12">
                 <div class="x_panel2">
                   <div class="x_title">
-                    <h2>Categories</h2>
-                    <ul class="nav navbar-right panel_toolbox">
-                      <li><a class="collapse-link"><i class="fa fa-chevron-up"></i></a>
-                      </li>
-                      <li class="dropdown">
-                        <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false"><i class="fa fa-wrench"></i></a>
-                        <ul class="dropdown-menu" role="menu">
-                          <li><a href="#">Settings 1</a>
-                          </li>
-                          <li><a href="#">Settings 2</a>
-                          </li>
-                        </ul>
-                      </li>
-                      <li><a class="close-link"><i class="fa fa-close"></i></a>
-                      </li>
-                    </ul>
+                    <h2><b>Categories</b></h2>
+
                     <div class="clearfix"></div>
                   </div>
                   <div class="x_content">
@@ -37,7 +23,7 @@
                           <th style="width: 1%">#</th>
                           <th style="width: 20%">Category ID</th>
                           <th style="width: 20%">Category Name</th>
-                          <th style="width: 20%">Action</th>
+                          <th style="width: 7%">Action</th>
 
                         </tr>
                       </thead>
@@ -53,7 +39,7 @@
 
                                 <td>
                                     <a href="EditCategory/<?php echo($data->id);?>" class="btn btn-info btn-xs"><i class="fa fa-pencil"></i> Edit </a>
-                                    <a href="deleteCategoryRecord/<?php echo($data->id);?>" class="btn btn-danger btn-xs"><i class="fa fa-trash-o"></i> Delete </a>
+                                    <a href="deleteCategoryRecord/<?php echo($data->id);?>" onclick="return confirm('Are you sure you want to delete this item?')" class="btn btn-danger btn-xs"><i class="fa fa-trash-o"> </i> Delete </a>
                                 </td>
 
                                 </tr>
